% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lead_entities.R
\name{load_lead_entities}
\alias{load_lead_entities}
\title{Load Lead Entities}
\usage{
load_lead_entities(download_files = FALSE)
}
\arguments{
\item{download_files}{A logical value indicating whether to save the lead entities data to the working directory as an ESRI shapefile.}
}
\value{
A simple features object containing the lead entities data.
}
\description{
This function reads lead entities data from an ESRI REST API as a simple features object and
saves it to the working directory as an ESRI shapefile if the download_files argument is set to TRUE.
}
\examples{
\dontrun{
lead_entities <- load_lead_entities()
head(lead_entities)
lead_entities_file <- load_lead_entities(download_files = TRUE)
}
}
